refactor: split auth view helpers by responsibility
